Add missing styles for some of the notes
WHATWG specs, including the Fetch spec, use a stylesheet different
from the W3C default which WinterCG uses. This WHATWG stylesheet
supports various types of notes and boxes with specific classes that
are not supported in the W3C default stylesheet.

This change adds styling for these classes, giving them a style
consistent with that of various types of notes in the W3C stylesheet.